drivers: clock: tms570: Fix RTI frequency
Previously VCLK was divided by 4 by default, as the TRM was
incorrectly interpreted. VCLK2 needs to be greater than or equal to,
whereas the previous implementation assumed it needed to be greater
than.

This fix, however, did not make sense as the RTI frequency was
correct when VCLK was divided by 4. The TRM does not state if
there is a divider for RTI when using VCLK as the source, but it
does state that other sources must ensure the frequency is 3 times
less than VCLK. Assuming the same rule applies when VCLK is used,
the RTI frequency becomes correct. This may be incorrect however,
and needs to be investigated further.