drivers: pinctrl: tms570: Write, not or, pin bit
Some pin bits need to be cleared in order for the write to have
any effect. The whole register is then cleared to 0, except for the
relevant bit. This can have the side effect of not being able to set
two bits that have no effect on eachother, so this may need to be
revisited later.